Federation prep: Refactor FindMember (#3928)
* Refactor FindMember

* Make error message generic

* Update src/GraphQL/Extensions/ObjectExtensions.cs


* Update src/GraphQL/Extensions/ObjectExtensions.cs


---------
